+++ Cargo.toml
@@ -11,4 +11,4 @@ pam="0.7.0"
 pam-sys="0.5.6"
 nix = "0.15.0"
 users="0.9.1"
-chvt = "0.1.3"
+chvt = "0.2.0"
